requirements.txt
backcall==0.2.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.2
cx-Oracle==8.1.0
cymem==2.0.5
Cython==0.29.22
decorator==4.4.2
docx2txt==0.8
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
httplib2==0.19.0
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.2
MarkupSafe==1.1.1
murmurhash==1.0.5
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
parso==0.8.1
pathlib==1.0.1
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
preshed==3.0.5
prompt-toolkit==3.0.16
prov==2.0.0
ptyprocess==0.7.0
pydantic==1.7.3
pydot==1.4.2
Pygments==2.8.1
PyMuPDF==1.18.9
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-pptx==0.6.18
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
spacy==3.0.3
spacy-legacy==3.0.1
sparse-dot-topn==0.2.9
srsly==2.4.0
thinc==8.0.1
threadpoolctl==2.1.0
tika==1.24
tqdm==4.58.0
traitlets==5.0.5
traits==6.2.0
typer==0.3.2
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
XlsxWriter==1.3.7
